third_party/WebKit/LayoutTests/hittesting/culled-inline.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<style>
#culled:hover {
  background-color: green;
}
.left {
  display: inline-block;
  transform: translateY(0px);
  border: 1px solid black;
}
.right {
  margin-left: 200px;
  border: 1px solid black;
}
</style>
<div id="container">
<span id="culled">
    <span class="left">Hello</span>
    <span class="right">World</span>
</span>
</div>
<script>
test(function(t)
{
    var rect = document.getElementById('culled').getBoundingClientRect();
    element = document.elementFromPoint(rect.left + rect.width/2, rect.top + rect.height/2);
    assert_equals(element.id, 'culled');
}, "Test that the space between two spans is returned when hittested");
</script>
